staging: fsl-dpaa2/eth: Updates to flow steering support
When flow steering was initially added MC did not support any
masking. Given this and the fact that both hashing and FS must
share a single key, the code was severely limited.

With the addition of masking, flow steering support in ethtool
was somewhat improved. We now take into account the flow type
and also support more key options.

MC currently supports a maximum number of 10 fields in a key.
If and when this limitation is addressed, more flow steering
options can be implemented.

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@nxp.com>