staging: fsl-dpaa2: Update logical FQs in netdevice
Currently, the netdevice is allocated with a default number of Rx/Tx
queues equal to CONFIG_NR_CPUS, meaning the maximum number of CPUs
supported by the current kernel. The correct number of queues is
reflected by the DPL settings.

Update the number of netdevice logical FQs after reading the DPNI
attributes.

Signed-off-by: Bogdan Purcareata <bogdan.purcareata@nxp.com>