staging: fsl-dpaa2: Flow affinity using queue mappings
The driver egress function chooses a FQ to send the frames based on the
current processor ID. The network stack itself sets a mapping field
in the skb based on many things - the default one being a hash on packet
fiels, which the current driver ignores.

One possibility is for the driver to adjust this mapping based on the
ingress FQ. This will further be considered by the network stack, and
there's no longer the need for an ndo_select_queue callback or for the
driver to choose his own queue on Tx.

The patch implements this change. In IPv4 forwarding scenarios, it
brings +4.8% perf in an affine flow scenario (16 flows) and +5.5% perf
in a 512 flow scenario (using hashing in the driver). Performance is
expressed in max throughput rate. Improvement is similar to implementing
a ndo_select_queue which sets the mapping based on the current processor
ID.

Signed-off-by: Bogdan Purcareata <bogdan.purcareata@nxp.com>